+++ ff/src/fields/mod.rs
@@ -153,6 +153,16 @@ pub trait Field:
     /// If the slice length != Self::extension_degree(), must return None.
     fn from_base_prime_field_elems(elems: &[Self::BasePrimeField]) -> Option<Self>;
 
+    /// Constructs a field element from a single base prime field elements.
+    /// ```
+    /// # use ark_ff::Field;
+    /// # use ark_test_curves::bls12_381::Fq as F;
+    /// # use ark_test_curves::bls12_381::Fq2 as F2;
+    /// # use ark_std::One;
+    /// assert_eq!(F2::from_base_prime_field(F::one()), F2::one());
+    /// ```
+    fn from_base_prime_field(elem: Self::BasePrimeField) -> Self;
+
     /// Returns `self + self`.
     #[must_use]
     fn double(&self) -> Self;