+++ ff/src/fields/models/cubic_extension.rs
@@ -134,11 +134,7 @@ impl<P: CubicExtConfig> CubicExtField<P> {
 
 impl<P: CubicExtConfig> Zero for CubicExtField<P> {
     fn zero() -> Self {
-        Self::new(
-            P::BaseField::zero(),
-            P::BaseField::zero(),
-            P::BaseField::zero(),
-        )
+        Self::new(P::BaseField::ZERO, P::BaseField::ZERO, P::BaseField::ZERO)
     }
 
     fn is_zero(&self) -> bool {
@@ -148,11 +144,7 @@ impl<P: CubicExtConfig> Zero for CubicExtField<P> {
 
 impl<P: CubicExtConfig> One for CubicExtField<P> {
     fn one() -> Self {
-        Self::new(
-            P::BaseField::one(),
-            P::BaseField::zero(),
-            P::BaseField::zero(),
-        )
+        Self::new(P::BaseField::ONE, P::BaseField::ZERO, P::BaseField::ZERO)
     }
 
     fn is_one(&self) -> bool {
@@ -170,6 +162,11 @@ impl<P: CubicExtConfig> Field for CubicExtField<P> {
 
     fn extension_degree() -> u64 {
         3 * P::BaseField::extension_degree()
+    }
+
+    fn from_base_prime_field(elem: Self::BasePrimeField) -> Self {
+        let fe = P::BaseField::from_base_prime_field(elem);
+        Self::new(fe, P::BaseField::ZERO, P::BaseField::ZERO)
     }
 
     fn to_base_prime_field_elements(&self) -> Self::BasePrimeFieldIter {
@@ -344,7 +341,7 @@ impl<P: CubicExtConfig> Zeroize for CubicExtField<P> {
 impl<P: CubicExtConfig> From<u128> for CubicExtField<P> {
     fn from(other: u128) -> Self {
         let fe: P::BaseField = other.into();
-        Self::new(fe, P::BaseField::zero(), P::BaseField::zero())
+        Self::new(fe, P::BaseField::ZERO, P::BaseField::ZERO)
     }
 }
 
@@ -363,7 +360,7 @@ impl<P: CubicExtConfig> From<i128> for CubicExtField<P> {
 impl<P: CubicExtConfig> From<u64> for CubicExtField<P> {
     fn from(other: u64) -> Self {
         let fe: P::BaseField = other.into();
-        Self::new(fe, P::BaseField::zero(), P::BaseField::zero())
+        Self::new(fe, P::BaseField::ZERO, P::BaseField::ZERO)
     }
 }
 
@@ -382,7 +379,7 @@ impl<P: CubicExtConfig> From<i64> for CubicExtField<P> {
 impl<P: CubicExtConfig> From<u32> for CubicExtField<P> {
     fn from(other: u32) -> Self {
         let fe: P::BaseField = other.into();
-        Self::new(fe, P::BaseField::zero(), P::BaseField::zero())
+        Self::new(fe, P::BaseField::ZERO, P::BaseField::ZERO)
     }
 }
 
@@ -401,7 +398,7 @@ impl<P: CubicExtConfig> From<i32> for CubicExtField<P> {
 impl<P: CubicExtConfig> From<u16> for CubicExtField<P> {
     fn from(other: u16) -> Self {
         let fe: P::BaseField = other.into();
-        Self::new(fe, P::BaseField::zero(), P::BaseField::zero())
+        Self::new(fe, P::BaseField::ZERO, P::BaseField::ZERO)
     }
 }
 
@@ -420,7 +417,7 @@ impl<P: CubicExtConfig> From<i16> for CubicExtField<P> {
 impl<P: CubicExtConfig> From<u8> for CubicExtField<P> {
     fn from(other: u8) -> Self {
         let fe: P::BaseField = other.into();
-        Self::new(fe, P::BaseField::zero(), P::BaseField::zero())
+        Self::new(fe, P::BaseField::ZERO, P::BaseField::ZERO)
     }
 }
 
@@ -440,8 +437,8 @@ impl<P: CubicExtConfig> From<bool> for CubicExtField<P> {
     fn from(other: bool) -> Self {
         Self::new(
             u8::from(other).into(),
-            P::BaseField::zero(),
-            P::BaseField::zero(),
+            P::BaseField::ZERO,
+            P::BaseField::ZERO,
         )
     }
 }
@@ -697,6 +694,22 @@ mod cube_ext_tests {
             let expected_2 = Fq2::new(random_coeffs[3], random_coeffs[4]);
             let expected = Fq6::new(expected_0, expected_1, expected_2);
             assert_eq!(actual, expected);
+        }
+    }
+
+    #[test]
+    fn test_from_base_prime_field_element() {
+        let ext_degree = Fq6::extension_degree() as usize;
+        let max_num_elems_to_test = 10;
+        for _ in 0..max_num_elems_to_test {
+            let mut random_coeffs = vec![Fq::zero(); ext_degree];
+            let random_coeff = Fq::rand(&mut test_rng());
+            let res = Fq6::from_base_prime_field(random_coeff);
+            random_coeffs[0] = random_coeff;
+            assert_eq!(
+                res,
+                Fq6::from_base_prime_field_elems(&random_coeffs).unwrap()
+            );
         }
     }
 }