+++ ff/src/fields/models/quadratic_extension.rs
@@ -244,6 +244,11 @@ impl<P: QuadExtConfig> Field for QuadExtField<P> {
         2 * P::BaseField::extension_degree()
     }
 
+    fn from_base_prime_field(elem: Self::BasePrimeField) -> Self {
+        let fe = P::BaseField::from_base_prime_field(elem);
+        Self::new(fe, P::BaseField::ZERO)
+    }
+
     fn to_base_prime_field_elements(&self) -> Self::BasePrimeFieldIter {
         self.c0
             .to_base_prime_field_elements()
@@ -410,12 +415,12 @@ where
             // so either a0 = sqrt(c0) or a1 = sqrt(c0/P::NONRESIDUE)
             if self.c0.legendre().is_qr() {
                 // either c0 is a valid sqrt in the base field
-                return self.c0.sqrt().map(|c0| Self::new(c0, P::BaseField::zero()));
+                return self.c0.sqrt().map(|c0| Self::new(c0, P::BaseField::ZERO));
             } else {
                 // or we need to compute sqrt(c0/P::NONRESIDUE)
                 return (self.c0.div(P::NONRESIDUE))
                     .sqrt()
-                    .map(|res| Self::new(P::BaseField::zero(), res));
+                    .map(|res| Self::new(P::BaseField::ZERO, res));
             }
         }
         // Try computing the square root
@@ -497,7 +502,7 @@ impl<P: QuadExtConfig> Zeroize for QuadExtField<P> {
 
 impl<P: QuadExtConfig> From<u128> for QuadExtField<P> {
     fn from(other: u128) -> Self {
-        Self::new(other.into(), P::BaseField::zero())
+        Self::new(other.into(), P::BaseField::ZERO)
     }
 }
 
@@ -515,7 +520,7 @@ impl<P: QuadExtConfig> From<i128> for QuadExtField<P> {
 
 impl<P: QuadExtConfig> From<u64> for QuadExtField<P> {
     fn from(other: u64) -> Self {
-        Self::new(other.into(), P::BaseField::zero())
+        Self::new(other.into(), P::BaseField::ZERO)
     }
 }
 
@@ -533,7 +538,7 @@ impl<P: QuadExtConfig> From<i64> for QuadExtField<P> {
 
 impl<P: QuadExtConfig> From<u32> for QuadExtField<P> {
     fn from(other: u32) -> Self {
-        Self::new(other.into(), P::BaseField::zero())
+        Self::new(other.into(), P::BaseField::ZERO)
     }
 }
 
@@ -551,7 +556,7 @@ impl<P: QuadExtConfig> From<i32> for QuadExtField<P> {
 
 impl<P: QuadExtConfig> From<u16> for QuadExtField<P> {
     fn from(other: u16) -> Self {
-        Self::new(other.into(), P::BaseField::zero())
+        Self::new(other.into(), P::BaseField::ZERO)
     }
 }
 
@@ -569,7 +574,7 @@ impl<P: QuadExtConfig> From<i16> for QuadExtField<P> {
 
 impl<P: QuadExtConfig> From<u8> for QuadExtField<P> {
     fn from(other: u8) -> Self {
-        Self::new(other.into(), P::BaseField::zero())
+        Self::new(other.into(), P::BaseField::ZERO)
     }
 }
 
@@ -587,7 +592,7 @@ impl<P: QuadExtConfig> From<i8> for QuadExtField<P> {
 
 impl<P: QuadExtConfig> From<bool> for QuadExtField<P> {
     fn from(other: bool) -> Self {
-        Self::new(u8::from(other).into(), P::BaseField::zero())
+        Self::new(u8::from(other).into(), P::BaseField::ZERO)
     }
 }
 
@@ -800,6 +805,22 @@ mod quad_ext_tests {
             let actual = Fq2::from_base_prime_field_elems(&random_coeffs).unwrap();
             let expected = Fq2::new(random_coeffs[0], random_coeffs[1]);
             assert_eq!(actual, expected);
+        }
+    }
+
+    #[test]
+    fn test_from_base_prime_field_element() {
+        let ext_degree = Fq2::extension_degree() as usize;
+        let max_num_elems_to_test = 10;
+        for _ in 0..max_num_elems_to_test {
+            let mut random_coeffs = vec![Fq::zero(); ext_degree];
+            let random_coeff = Fq::rand(&mut test_rng());
+            let res = Fq2::from_base_prime_field(random_coeff);
+            random_coeffs[0] = random_coeff;
+            assert_eq!(
+                res,
+                Fq2::from_base_prime_field_elems(&random_coeffs).unwrap()
+            );
         }
     }
 }